Standardization

+++ Pod/Classes/GDGBelongsToRelation.m
@@ -1,66 +1,67 @@
 //
 //  GDGBelongsToRelation.m
-//  Pods
+//  GoldDigger
 //
 //  Created by Pietro Caselani on 1/26/16.
-//
 //
 
 #import "GDGBelongsToRelation.h"
 
 #import <GoldDigger/GDGConditionBuilder+EntityQuery.h>
+#import <ObjectiveSugar/ObjectiveSugar.h>
 #import "GDGEntityQuery.h"
 #import "GDGEntitySettings.h"
 #import "GDGTableSource.h"
-#import <ObjectiveSugar/ObjectiveSugar.h>
 
 @implementation GDGBelongsToRelation
 
-- (void)fill:(NSArray<GDGEntity*>*)entities withProperties:(NSArray<NSString*>*)properties
+- (void)fill:(NSArray<GDGEntity *> *)entities withProperties:(NSArray<NSString *> *)properties
 {
-	NSArray<NSNumber*>* ids = [entities map:^id(id object) {
+	NSArray<NSNumber *> *ids = [entities map:^id(id object) {
 		return [object valueForKey:self.foreignProperty];
 	}];
-	
-	NSMutableDictionary<NSNumber*, NSMutableArray<GDGEntity*>*>* relationEntitiesDictionary = [[NSMutableDictionary alloc] initWithCapacity:entities.count];
-	for (int i = 0; i < entities.count; i++)
+
+	NSMutableDictionary<NSNumber *, NSMutableArray<GDGEntity *> *> *relationEntitiesDictionary = [[NSMutableDictionary alloc] initWithCapacity:entities.count];
+	for (NSUInteger i = 0; i < entities.count; i++)
 	{
-		GDGEntity* entity = entities[i];
-		NSNumber* foreignId = ids[i];
-		
-		NSMutableArray<GDGEntity*>* mutableEntities = relationEntitiesDictionary[foreignId];
+		GDGEntity *entity = entities[i];
+		NSNumber *foreignId = ids[i];
+
+		NSMutableArray<GDGEntity *> *mutableEntities = relationEntitiesDictionary[foreignId];
 		if (mutableEntities == nil)
 		{
 			mutableEntities = [NSMutableArray array];
 			relationEntitiesDictionary[foreignId] = mutableEntities;
 		}
-		
+
 		[mutableEntities addObject:entity];
 	}
-	
-	GDGEntityQuery* query = self.relatedManager.query.select(properties)
-		.where(^(GDGConditionBuilder *builder) { builder.prop(@"id").inList(ids); });
-	
+
+	GDGEntityQuery *query = self.relatedManager.query.select(properties)
+			.where(^(GDGConditionBuilder *builder) {
+				builder.prop(@"id").inList(ids);
+			});
+
 	if (self.condition)
 		query.where(^(GDGConditionBuilder *builder) {
 			builder.and.cat(self.condition);
 		});
-	
-	for (GDGEntity* relatedEntity in query.array)
+
+	for (GDGEntity *relatedEntity in query.array)
 	{
-		NSMutableArray<GDGEntity*>* mutableEntities = relationEntitiesDictionary[@([relatedEntity id])];
-		for (GDGEntity* entity in mutableEntities)
+		NSMutableArray<GDGEntity *> *mutableEntities = relationEntitiesDictionary[@([relatedEntity id])];
+		for (GDGEntity *entity in mutableEntities)
 			[entity setValue:relatedEntity forKey:self.name];
 	}
 }
 
-- (NSString*)joinCondition
+- (NSString *)joinCondition
 {
 	NSMutableString *condition = [[NSMutableString alloc] initWithString:self.manager.settings.tableSource.alias];
-	
+
 	[condition appendString:@".id"];
 	[condition appendFormat:@" = %@.%@", self.relatedManager.settings.tableSource.alias, self.foreignProperty];
-	
+
 	return [NSString stringWithString:condition];
 }
